Move model to a separate assembly
