+++ nunit-gui.sln
@@ -24,6 +24,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		NuGet.config = NuGet.config
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "nunit.gui", "src\nunit.gui\nunit.gui.csproj", "{E978A8D7-E4A4-4DE9-94DF-D82BD764936B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -46,6 +48,10 @@ Global
 		{2E368281-3BA8-4050-B05E-0E0E43F8F446}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2E368281-3BA8-4050-B05E-0E0E43F8F446}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2E368281-3BA8-4050-B05E-0E0E43F8F446}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E978A8D7-E4A4-4DE9-94DF-D82BD764936B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E978A8D7-E4A4-4DE9-94DF-D82BD764936B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E978A8D7-E4A4-4DE9-94DF-D82BD764936B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E978A8D7-E4A4-4DE9-94DF-D82BD764936B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
