+++ src/nunit-gui/Views/SettingsPages/GuiSettingsPage.cs
@@ -25,7 +25,6 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Drawing;
-using System.Data;
 using System.Text;
 using System.Windows.Forms;
 using NUnit.Gui.Model;
