+++ src/nunit.gui/Model/ResultNode.cs
@@ -21,12 +21,8 @@
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 // ***********************************************************************
 
-using System;
-using System.Collections.Generic;
 using System.Globalization;
-using System.Text;
 using System.Xml;
-using NUnit.Engine;
 
 namespace NUnit.Gui.Model
 {
