+++ src/nunit.gui/Model/ResultState.cs
