+++ src/nunit.gui/Model/RunState.cs
