+++ src/nunit.gui/Model/Settings/GuiSettings.cs
