+++ src/nunit.gui/Model/Settings/SettingsWrapper.cs
@@ -21,9 +21,6 @@
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 // ***********************************************************************
 
-using System.Drawing;
-using System.Windows.Forms;
-
 namespace NUnit.Gui.Model.Settings
 {
     using Engine;
