+++ src/nunit.gui/Model/TestEventArgs.cs
@@ -22,7 +22,6 @@
 // ***********************************************************************
 
 using System;
-using System.Xml;
 
 namespace NUnit.Gui.Model
 {
