+++ src/nunit.gui/Model/TestModel.cs
@@ -23,7 +23,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Threading;
 using System.Windows.Forms;
 using System.Xml;
 using NUnit.Engine;
