+++ src/nunit.gui/Model/TestSelection.cs
@@ -22,11 +22,7 @@
 // ***********************************************************************
 
 using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.Text;
-using System.Xml;
-using NUnit.Engine;
 
 namespace NUnit.Gui.Model
 {
