+++ src/nunit.gui/Model/TestStatus.cs
