+++ src/nunit.gui/Options.cs
