+++ src/nunit.gui/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="NUnit.Engine" version="3.6.1" targetFramework="net20" />
+</packages>
